refactoring of array plotting; improved functionality
